chore: move tests into logicals... its only logical.
